# Check implied weight validity when computing normalizing constant