In the future we might allow users to provide custom implied weights, so this seems like a natural place to check that implied weights are valid (i.e. that they reweight edges that exist and don't change the sign of an edge weight).