@oscarhiggott 2025-08-06
The extra checks don't seem to impact performance, here is the dem loading from `pymatching_perf`:

Before:

```
[...................*|....................]  47 ms (vs  35 ms) (200 loads/s) Load_dem_r11_d11_p100
[..................*<|....................] 390 ms (vs 280 ms) ( 25 loads/s) Load_dem_r21_d21_p100
[.................*<<|....................]  77 ms (vs  35 ms) (120 loads/s) Load_dem_r11_d11_p100_correlations
[................*<<<|....................] 660 ms (vs 280 ms) ( 15 loads/s) Load_dem_r21_d21_p100_correlations
```

After:

```
[...................*|....................]  47 ms (vs  35 ms) (200 loads/s) Load_dem_r11_d11_p100
[..................*<|....................] 400 ms (vs 280 ms) ( 24 loads/s) Load_dem_r21_d21_p100
[.................*<<|....................]  71 ms (vs  35 ms) (140 loads/s) Load_dem_r11_d11_p100_correlations
[................*<<<|....................] 630 ms (vs 280 ms) ( 15 loads/s) Load_dem_r21_d21_p100_correlations
```